requirements.txt
attrs==20.2.0
coverage==5.3
flake8==3.8.4
future==0.18.2
importlib-metadata==2.0.0
iniconfig==1.1.1
iso8601==0.1.13
mccabe==0.6.1
numpy==1.19.3
packaging==20.4
paho-mqtt==1.5.1
pandas==1.1.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pyserial==3.4
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
RPi.GPIO==0.7.0
six==1.15.0
toml==0.10.1
zipp==3.4.0
